Add dragging handle to window actions
