+++ client/Page/Admin/Chat/Partials/ChatTopBar.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useRef } from 'react';
+import React from 'react';
 import { cn } from '@/lib/utils';
 import AddIcon from '@material-design-icons/svg/outlined/add.svg?react';
 import HistoryIcon from '@material-design-icons/svg/outlined/history.svg?react';
@@ -6,14 +6,11 @@ import SettingsIcon from '@material-design-icons/svg/outlined/settings.svg?react
 import AccountIcon from '@material-design-icons/svg/outlined/account_circle.svg?react';
 import { Button } from '@/Components/ui/button';
 import { Badge } from '@/Components/ui/badge';
-import { useClientSettings } from '@/Providers/ClientSettingsProvider';
 import Logo from '@/Components/Logo';
 import { useChat } from '@/Providers/ChatProvider';
 
 export default function ChatTopBar() {
-  const { setSettings } = useClientSettings();
   const { openChatOverlay } = useChat();
-  const topBarRef = useRef(null);
 
   function onUpgradeClick(e: React.FormEvent) {
     e.preventDefault();
@@ -40,85 +37,11 @@ export default function ChatTopBar() {
     openChatOverlay('Profile');
   }
 
-  function startDrag(e) {
-    e.preventDefault();
-    const chatWindow = document.getElementById('awp-chat');
-    if (chatWindow.classList.contains('maximized')) {
-      return;
-    }
-    const bodyElement = document.getElementsByTagName('body')[0];
-    const containerElement = document.getElementById('wpbody');
-    const containerCoords = containerElement.getBoundingClientRect();
-    const initialMousePositionX = e.clientX;
-    const initialMousePositionY = e.clientY;
-    const computedStyle = window.getComputedStyle(chatWindow);
-    const matrix = new DOMMatrixReadOnly(computedStyle.transform);
-
-    const initialPositionX = matrix.m41;
-    const initialPositionY = matrix.m42;
-    let lastPositionX = matrix.m41;
-    let lastPositionY = matrix.m42;
-
-    let startPosX = e.clientX - e.target.getBoundingClientRect().left;
-    let startPosY = e.clientY - e.target.getBoundingClientRect().top;
-
-    const disableDrag = (e) => {
-      setSettings({
-        x: lastPositionX,
-        y: lastPositionY,
-      });
-      bodyElement.removeEventListener('mousemove', handleDrag);
-      bodyElement.onmouseup = null;
-    };
-
-    const drag = (target, x, y) => {
-      const currentPositionMatrix = new DOMMatrixReadOnly(computedStyle.transform);
-
-      // check if any of the new bounding rects are out of bounds
-      const newPositionLeft = x - startPosX;
-      const newPositionTop = y - startPosY;
-      const leftOutOfBounds = newPositionLeft < containerCoords.left;
-      const topOutOfBounds = newPositionTop < containerCoords.top;
-
-      let newCoordsX = currentPositionMatrix.m41;
-      let newCoordsY = currentPositionMatrix.m42;
-
-      if (!leftOutOfBounds) {
-        newCoordsX = x - initialMousePositionX + initialPositionX;
-      }
-
-      if (!topOutOfBounds) {
-        newCoordsY = y - initialMousePositionY + initialPositionY;
-      }
-
-      target.style.transform = `translate(${newCoordsX}px, ${newCoordsY}px)`;
-      lastPositionX = newCoordsX;
-      lastPositionY = newCoordsY;
-    };
-
-    const handleDrag = (e) => {
-      drag(chatWindow, e.clientX, e.clientY);
-    };
-
-    bodyElement.addEventListener('mousemove', handleDrag);
-    bodyElement.addEventListener('mouseup', disableDrag);
-  }
-
-  useEffect(() => {
-    const element = topBarRef.current;
-    element.addEventListener('mousedown', startDrag);
-
-    return () => {
-      element.removeEventListener('mousedown', startDrag);
-    };
-  }, []);
-
   return (
     <div
-      ref={topBarRef}
       className={cn(
         'py-2 px-2 border-b border-b-brand-gray-25',
-        'flex justify-between cursor-move',
+        'flex justify-between',
       )}
     >
       <div className="flex items-center gap-2">
